refactor: extract helpers from App
